Converted to uv/pipx installable package
